bug fixes with cookies and api calls

+++ src/app/api/v1/signout/route.js
@@ -4,17 +4,26 @@ import { cookies } from 'next/headers';
 export async function DELETE(req){
     try{
         const {db} = await connectToDatabase();
-        const body = await req.json();
-
-        // log api call
-        const apiCallsLog = await db.collection('apiCalls').updateOne({email: body.email}, {$inc: {signout: 1}});
         
         const cookieStore = await cookies();
         const sessionToken = cookieStore.get('sessionToken')?.value;
 
+        //check sessionToken from cookie
         if(!sessionToken){
             return new Response(JSON.stringify({error:"No session token found"}), {status: 400});
         }
+        
+        //find session in db
+        const session = await db.collection('sessions').findOne({ token: sessionToken });
+
+        if (!session) {
+            return new Response(JSON.stringify({ error: "Session not found" }), { status: 404 });
+        }        
+
+        const email = session.email;
+
+        // log api call
+        const apiCallsLog = await db.collection('apiCalls').updateOne({email: email}, {$inc: {signout: 1}});
 
         await db.collection('sessions').deleteOne({token:sessionToken});
